move providers tests into a separate test file
